requirements.txt
astroid==2.4.2
autopep8==1.4.4
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.42.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==1.10.1
django-dotenv==1.4.2
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==8.4.0
pycodestyle==2.5.0
pylint==2.5.3
python-decouple==3.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2018.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
stripe==2.49.0
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
django-hitcount
